+++ js/util.js
@@ -1,0 +1,25 @@
+// вспомогательныe функции
+// Функция взята из интернета и доработана
+// Источник - https://github.com/you-dont-need/You-Dont-Need-Lodash-Underscore#_random
+
+function getRandomPositiveInteger (a, b) {
+  if (a < 0 || b < 0) {
+    return NaN;
+  }
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+}
+
+// getRandomPositiveInteger ();
+
+// Функция для проверки максимальной длины строки.
+
+function verifyMaxLegthStr(str, maxlength) {
+  return str.length <= maxlength;
+}
+
+verifyMaxLegthStr ('Maksar', 10);
+
+export {getRandomPositiveInteger};
